RPC snapshot correction
After filing an issue on github about snapshots failing to create, I was made aware that I was creating them too soon.
A snapshot is now made ever 3000 blocks.
It was recommended to do it only every 250k.
https://github.com/cosmos/cosmos-sdk/issues/9686#issuecomment-1064271936

With this correction, the snapshot's are taken every 3000 blocks, and 100 are kept.  So 300k blocks can be restored in the event of some unforeseen circumstance.