Link issue
Landing page for the docs had a bad link I spotted.  Finally found where it was stored and corrected it.  Currently, the live page will give a 404 redirect when clicking on the link.  I am not sure why it's not routing properly, as my local copy works fine.
Maybe this isn't the 'proper' fix but it will resolve the issue.