RPC Snapshot interval increased
I decided to increase the snapshot interval on my RPC server from 3000 to 375K blocks.
I did this to prevent state sync abuse.  This equates to a little over of 1 months worth of blocks.

Validators should maintain a min of 2/3 unbonding time of blocks.  This will help ensure that they do that.